fix(theta): refine jaccard API
