+++ datasketches/src/common/mod.rs
@@ -24,5 +24,7 @@ pub use self::resize::ResizeFactor;
 
 #[cfg(feature = "theta")]
 pub(crate) mod binomial_bounds;
+#[cfg(feature = "theta")]
+pub(crate) mod bounds_binomial_proportions;
 #[cfg(feature = "cpc")]
 pub(crate) mod inv_pow2_table;
