+++ datasketches/src/theta/mod.rs
@@ -49,7 +49,6 @@ mod union;
 
 pub use self::intersection::ThetaIntersection;
 pub use self::jaccard_similarity::JaccardSimilarity;
-pub use self::jaccard_similarity::ThetaJaccardSimilarity;
 pub use self::sketch::CompactThetaSketch;
 pub use self::sketch::ThetaSketch;
 pub use self::sketch::ThetaSketchBuilder;
